Refactor SequenceFeatureDetails
