+++ packages/core/BaseFeatureWidget/SequenceFeatureDetails/SequencePanel.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { defaultCodonTable, generateCodonTable, revcom } from '../util'
+import { defaultCodonTable, generateCodonTable, revcom } from '../../util'
 import {
   ParentFeat,
   SeqState,
@@ -8,8 +8,11 @@ import {
   calculateUTRs2,
   dedupe,
   revlist,
-} from './util'
-import { GenecDNA, GeneProtein, GeneCDS, Genomic } from './SequenceBox'
+} from '../util'
+import CDNASequence from './CDNASequence'
+import ProteinSequence from './ProteinSequence'
+import GenomicSequence from './GenomicSequence'
+import CDSSequence from './CDSSequence'
 
 interface SeqPanelProps {
   sequence: SeqState
@@ -99,13 +102,17 @@ const SeqPanel = React.forwardRef<HTMLDivElement, SeqPanelProps>(function (
         </span>
         <br />
         {mode === 'genomic' ? (
-          <Genomic sequence={seq} />
-        ) : mode === 'genomic_sequence_updown' ? (
-          <Genomic sequence={seq} upstream={upstream} downstream={downstream} />
+          <GenomicSequence sequence={seq} />
+        ) : mode === 'genomic_sequence_updownstream' ? (
+          <GenomicSequence
+            sequence={seq}
+            upstream={upstream}
+            downstream={downstream}
+          />
         ) : mode === 'cds' ? (
-          <GeneCDS cds={cds} sequence={seq} />
+          <CDSSequence cds={cds} sequence={seq} />
         ) : mode === 'cdna' ? (
-          <GenecDNA
+          <CDNASequence
             exons={exons}
             cds={cds}
             utr={utr}
@@ -113,9 +120,9 @@ const SeqPanel = React.forwardRef<HTMLDivElement, SeqPanelProps>(function (
             intronBp={intronBp}
           />
         ) : mode === 'protein' ? (
-          <GeneProtein cds={cds} codonTable={codonTable} sequence={seq} />
+          <ProteinSequence cds={cds} codonTable={codonTable} sequence={seq} />
         ) : mode === 'gene' ? (
-          <GenecDNA
+          <CDNASequence
             exons={exons}
             cds={cds}
             utr={utr}
@@ -124,7 +131,7 @@ const SeqPanel = React.forwardRef<HTMLDivElement, SeqPanelProps>(function (
             intronBp={intronBp}
           />
         ) : mode === 'gene_collapsed_intron' ? (
-          <GenecDNA
+          <CDNASequence
             exons={exons}
             cds={cds}
             sequence={seq}
@@ -134,7 +141,7 @@ const SeqPanel = React.forwardRef<HTMLDivElement, SeqPanelProps>(function (
             intronBp={intronBp}
           />
         ) : mode === 'gene_updownstream' ? (
-          <GenecDNA
+          <CDNASequence
             exons={exons}
             cds={cds}
             sequence={seq}
@@ -145,7 +152,7 @@ const SeqPanel = React.forwardRef<HTMLDivElement, SeqPanelProps>(function (
             intronBp={intronBp}
           />
         ) : mode === 'gene_updownstream_collapsed_intron' ? (
-          <GenecDNA
+          <CDNASequence
             exons={exons}
             cds={cds}
             sequence={seq}
