+++ packages/core/BaseFeatureWidget/SequenceFeatureDetails/test_data/DLGAP3.js
