+++ packages/core/BaseFeatureWidget/SequenceFeatureDetails/test_data/DLGAP3_pep.fa
