+++ packages/core/BaseFeatureWidget/SequenceFeatureDetails/test_data/NCDN.js
