+++ packages/core/BaseFeatureWidget/SequenceFeatureDetails/test_data/NCDN_dna.fa
