+++ packages/core/BaseFeatureWidget/SequenceFeatureDetails/test_data/NCDN_upstream_dna.fa
