+++ packages/core/BaseFeatureWidget/SequenceFeatureDetails/test_data/volvox.fa
